+++ functions/src/index.ts
@@ -5,3 +5,4 @@ export {migrateHouseholdClaims, migrateHouseholdClaimsHttp} from "./migrateHouse
 export {getNutritionData} from "./nutrition";
 export {resetWeeklyUsageLimits, resetUsageLimitsNow} from "./resetUsageLimits";
 export {verifyPurchase} from "./verifyPurchase";
+export {deleteAccount} from "./deleteAccount";